Add a sharable AsyncIterator
This commit should've come earlier than the last few. When native
AsyncGenerators are `next`'ed while they're still generating values,
they throw a "Generator is already running". (Known since HHReactor).
Instead of breaking out a full buffered `Producer` like HHReactor, I'm
fairly confident this is along the lines of the most minimal way to
share an AsyncGenerator. Interface is for enforcing sharing, but it's
a weak strategy that's subject to disappearing later.